requirements.txt
Flask==3.0.0
pytest==7.4.3
pytest-cov==4.1.0
gunicorn==21.2.0
python-dotenv==1.0.0
pyyaml==6.0.1

